+++ CHANGE_LOG.md
@@ -1,3 +1,9 @@
+# 1.1.0
+
+* Expose manually locking and unlocking processes.
+
+* Allow passing in a redis connection in the constructor.
+
 # 1.0.1
 
 * Use Lua script to avoid race conditions and ensure no extra processes slip through.